Move config for clients to root-app
Trenger dette for å hente inn PdlClient i identhendelse-servicen
