+++ src/test/kotlin/testhelper/TestApiModule.kt
@@ -3,6 +3,8 @@ package testhelper
 import io.ktor.server.application.*
 import no.nav.syfo.application.api.apiModule
 import no.nav.syfo.application.cache.RedisStore
+import no.nav.syfo.client.azuread.AzureAdClient
+import no.nav.syfo.client.pdl.PdlClient
 
 fun Application.testApiModule(
     externalMockEnvironment: ExternalMockEnvironment,
@@ -10,12 +12,23 @@ fun Application.testApiModule(
     val redisStore = RedisStore(
         redisEnvironment = externalMockEnvironment.environment.redis,
     )
+    val azureAdClient = AzureAdClient(
+        azureEnviroment = externalMockEnvironment.environment.azure,
+        redisStore = redisStore,
+    )
+    val pdlClient = PdlClient(
+        azureAdClient = azureAdClient,
+        clientEnvironment = externalMockEnvironment.environment.clients.pdl,
+        redisStore = redisStore,
+    )
     this.apiModule(
         applicationState = externalMockEnvironment.applicationState,
+        azureAdClient = azureAdClient,
         database = externalMockEnvironment.database,
         environment = externalMockEnvironment.environment,
         wellKnownInternalAzureAD = externalMockEnvironment.wellKnownInternalAzureAD,
         wellKnownSelvbetjening = externalMockEnvironment.wellKnownSelvbetjening,
         redisStore = redisStore,
+        pdlClient = pdlClient,
     )
 }
